% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcorrplot.R
\name{catcorrplot}
\alias{catcorrplot}
\title{Create correlation plot}
\usage{
catcorrplot(df)
}
\arguments{
\item{df}{dataframe containing categorical vectors}
}
\description{
Final nested function of this package after
crmvtbl_condense
}
\details{
This function takes a dataframe and outputs a Cramer's V proxy
correlation plot
}
\examples{
catcorrplot()
}
