% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmv.R
\name{crmv}
\alias{crmv}
\title{Cramer's V of fields}
\usage{
crmv(varb1, varb2)
}
\arguments{
\item{varb1}{First vector}

\item{varb2}{Second vector}
}
\description{
Used https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
to create this package.
}
\details{
This function takes two vectors and outputs their Cramer's V,
which is a proxy for correlation.
}
\examples{
crmv()
}
