% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmvtbl.R
\name{crmvtbl}
\alias{crmvtbl}
\title{CV Table Build}
\usage{
crmvtbl(df)
}
\arguments{
\item{df}{dataframe containing categorical variables}
}
\description{
Third nested function after onevarcv of this package
}
\details{
This function takes a dataframe and outputs the Cramer's V
for all pairs of variables.
}
\examples{
crmvtbl()
}
