% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmvtbl_condense.R
\name{crmvtbl_condense}
\alias{crmvtbl_condense}
\title{Helper Table Build for CatCorrPlot}
\usage{
crmvtbl_condense(df)
}
\arguments{
\item{df}{dataframe containing categorical variables}
}
\description{
Fourth nested function after crmvtbl of this package
}
\details{
This function takes a dataframe and makes a new table
compatible to be fed to CatCorrPlot
}
\examples{
crmvtbl_condense()
}
