% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevarcv.R
\name{onevarcv}
\alias{onevarcv}
\title{Cramer's v of 1 field with dataframe}
\usage{
onevarcv(var, df)
}
\arguments{
\item{var}{vector that cramer's v will be calculated with}

\item{df}{dataframe containing other vectors which will have Cramer's v calculated with var}
}
\description{
Second nested function after crmv of this package
}
\details{
This function takes one vector and calculates the
Cramer's V for each other vector in a dataframe.
}
\examples{
onevarcv()
}
