% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcorr.R
\name{catCorr}
\alias{catCorr}
\title{catCorr}
\usage{
catCorr(x = NULL, y = NULL, bias_correction = TRUE, type = "both")
}
\arguments{
\item{x}{Either a vector or data.frame to analyze.}

\item{y}{If \code{x} is a vector, the \code{y} is the vector to compare it to.}

\item{bias_correction}{Logical, if TRUE (default), use the bias correction in
Cramer's V}

\item{type}{Which statistics to produce.  Options are \code{'tschuprow'},
\code{'tschu'}, \code{'t'} for Tschuprow's T statistic, \code{'cramer'},
\code{'c'}, \code{'v'} for Cramer's V statistic and p-values, and
\code{'both'} or \code{'b'} for both options.}
}
\value{
Cramer's V (\code{cramer_V}), Tschuprow's T (\code{tschu_T}), and/or
p-value(s) (\code{pval}).
}
\description{
Produce metrics of correlation among categorical variables.
}

